Internal changes and optimization  - Made add new task work with correct position updates  - refactoring of states and events and their uses in UI and view model.  - Other lots of refactoring and refinement.

+++ app/src/main/java/com/app/routineturboa/data/repository/TaskOperationResult.kt
@@ -2,6 +2,6 @@ package com.app.routineturboa.data.repository
 
 data class TaskOperationResult(
     val success: Boolean,
-    val newTaskId: Long? = null
+    val newTaskId: Int? = null
 )
